.scripts/hash_collision_checker/requirements.txt
# dependencies for check_murmur.py
PyYAML
mmh3
